Convert to DSD model.
