+++ openchrom/plugins/net.openchrom.wsd.converter.supplier.abif/src/net/openchrom/wsd/converter/supplier/abif/core/ChromatogramImportConverter.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2016, 2025 Matthias Mailänder, Philip Wenig.
+ * Copyright (c) 2016, 2026 Lablicate GmbH.
  *
  * This program and the accompanying materials are made
  * available under the terms of the Eclipse Public License 2.0
@@ -18,30 +18,30 @@ import java.io.IOException;
 
 import org.eclipse.chemclipse.converter.chromatogram.AbstractChromatogramImportConverter;
 import org.eclipse.chemclipse.converter.l10n.ConverterMessages;
+import org.eclipse.chemclipse.dsd.converter.io.IChromatogramDSDReader;
+import org.eclipse.chemclipse.dsd.model.core.IChromatogramDSD;
 import org.eclipse.chemclipse.logging.core.Logger;
 import org.eclipse.chemclipse.model.core.IChromatogramOverview;
 import org.eclipse.chemclipse.processing.core.IProcessingInfo;
-import org.eclipse.chemclipse.wsd.converter.io.IChromatogramWSDReader;
-import org.eclipse.chemclipse.wsd.model.core.IChromatogramWSD;
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.osgi.util.NLS;
 
 import net.openchrom.wsd.converter.supplier.abif.io.ChromatogramReader;
 
-public class ChromatogramImportConverter extends AbstractChromatogramImportConverter<IChromatogramWSD> {
+public class ChromatogramImportConverter extends AbstractChromatogramImportConverter<IChromatogramDSD> {
 
 	private static final Logger logger = Logger.getLogger(ChromatogramImportConverter.class);
 	private static final String DESCRIPTION = "ABIF Import Converter";
 
 	@Override
-	public IProcessingInfo<IChromatogramWSD> convert(File file, IProgressMonitor monitor) {
+	public IProcessingInfo<IChromatogramDSD> convert(File file, IProgressMonitor monitor) {
 
-		IProcessingInfo<IChromatogramWSD> processingInfo = super.validate(file);
+		IProcessingInfo<IChromatogramDSD> processingInfo = super.validate(file);
 		if(!processingInfo.hasErrorMessages()) {
-			IChromatogramWSDReader reader = new ChromatogramReader();
+			IChromatogramDSDReader reader = new ChromatogramReader();
 			monitor.subTask(ConverterMessages.importChromatogram);
 			try {
-				IChromatogramWSD chromatogram = reader.read(file, monitor);
+				IChromatogramDSD chromatogram = reader.read(file, monitor);
 				processingInfo.setProcessingResult(chromatogram);
 			} catch(IOException e) {
 				logger.warn(e);
@@ -56,7 +56,7 @@ public class ChromatogramImportConverter extends AbstractChromatogramImportConve
 
 		IProcessingInfo<IChromatogramOverview> processingInfo = super.validate(file);
 		if(!processingInfo.hasErrorMessages()) {
-			IChromatogramWSDReader reader = new ChromatogramReader();
+			IChromatogramDSDReader reader = new ChromatogramReader();
 			monitor.subTask(ConverterMessages.importChromatogramOverview);
 			try {
 				IChromatogramOverview chromatogramOverview = reader.readOverview(file, monitor);
