+++ openchrom/plugins/net.openchrom.wsd.converter.supplier.abif/src/net/openchrom/wsd/converter/supplier/abif/internal/support/ChromatogramArrayReader.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2016, 2025 Matthias Mailänder, Philip Wenig.
+ * Copyright (c) 2016, 2026 Lablicate GmbH.
  *
  * This program and the accompanying materials are made
  * available under the terms of the Eclipse Public License 2.0
@@ -21,6 +21,7 @@ import org.eclipse.chemclipse.converter.io.support.AbstractArrayReader;
 public class ChromatogramArrayReader extends AbstractArrayReader implements IChromatogramArrayReader {
 
 	public ChromatogramArrayReader(File file) throws IOException {
+
 		super(file);
 	}
 }
