+++ openchrom/plugins/net.openchrom.wsd.converter.supplier.abif/src/net/openchrom/wsd/converter/supplier/abif/model/IVendorChromatogram.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2016, 2025 Matthias Mailänder, Philip Wenig.
+ * Copyright (c) 2016, 2026 Lablicate GmbH.
  *
  * This program and the accompanying materials are made
  * available under the terms of the Eclipse Public License 2.0
@@ -13,9 +13,9 @@
  *******************************************************************************/
 package net.openchrom.wsd.converter.supplier.abif.model;
 
-import org.eclipse.chemclipse.wsd.model.core.IChromatogramWSD;
+import org.eclipse.chemclipse.dsd.model.core.IChromatogramDSD;
 
-public interface IVendorChromatogram extends IChromatogramWSD {
+public interface IVendorChromatogram extends IChromatogramDSD {
 
 	short getVersion();
