+++ openchrom/tests/net.openchrom.wsd.converter.supplier.abif.fragment.test/src/net/openchrom/dsd/converter/supplier/abif/io/EmptyFile_ITest.java
@@ -9,17 +9,16 @@
  * 
  * Contributors:
  * Matthias Mailänder - initial API and implementation
- * Philip Wenig - initial API and implementation
  *******************************************************************************/
-package net.openchrom.wsd.converter.supplier.abif.io;
+package net.openchrom.dsd.converter.supplier.abif.io;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 import java.io.File;
 
+import org.eclipse.chemclipse.dsd.model.core.IChromatogramDSD;
 import org.eclipse.chemclipse.processing.core.IProcessingInfo;
-import org.eclipse.chemclipse.wsd.model.core.IChromatogramWSD;
 import org.eclipse.core.runtime.NullProgressMonitor;
 import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Order;
@@ -32,24 +31,30 @@ import net.openchrom.wsd.converter.supplier.abif.core.ChromatogramImportConverte
 
 @TestInstance(Lifecycle.PER_CLASS)
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class AbiPrism3730GeneticAnalyser_ITest {
+public class EmptyFile_ITest {
 
-	private IChromatogramWSD chromatogram;
+	private IChromatogramDSD chromatogram;
 
 	@Test
 	@Order(1)
 	public void testImport() {
 
-		File fileImport = new File("testdata/files/import/3730.ab1");
+		File fileImport = new File("testdata/files/import/empty.ab1");
 		ChromatogramImportConverter importConverter = new ChromatogramImportConverter();
-		IProcessingInfo<IChromatogramWSD> processingInfo = importConverter.convert(fileImport, new NullProgressMonitor());
+		IProcessingInfo<IChromatogramDSD> processingInfo = importConverter.convert(fileImport, new NullProgressMonitor());
 		chromatogram = processingInfo.getProcessingResult();
 		assertNotNull(chromatogram);
 	}
 
 	@Test
-	public void testScans() {
+	public void testEmptySequence() {
 
-		assertEquals(16302, chromatogram.getNumberOfScans());
+		assertEquals("NNNNN", chromatogram.getNucleotideSequence());
+	}
+
+	@Test
+	public void testSampleName() {
+
+		assertEquals("226041_C-ME-19_pCAGseqF", chromatogram.getSampleName());
 	}
 }
