--- openchrom/tests/net.openchrom.wsd.converter.supplier.abif.fragment.test/src/net/openchrom/wsd/converter/supplier/abif/io/AbiPrism3100GeneticAnalyser_ITest.java
@@ -1,57 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2016, 2026 Matthias Mailänder, Philip Wenig.
- *
- * This program and the accompanying materials are made
- * available under the terms of the Eclipse Public License 2.0
- * which is available at https://www.eclipse.org/legal/epl-2.0/
- *
- * SPDX-License-Identifier: EPL-2.0
- * 
- * Contributors:
- * Matthias Mailänder - initial API and implementation
- * Philip Wenig - initial API and implementation
- *******************************************************************************/
-package net.openchrom.wsd.converter.supplier.abif.io;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-
-import java.io.File;
-
-import org.eclipse.chemclipse.processing.core.IProcessingInfo;
-import org.eclipse.chemclipse.wsd.model.core.IChromatogramWSD;
-import org.eclipse.chemclipse.wsd.model.core.IScanWSD;
-import org.eclipse.core.runtime.NullProgressMonitor;
-import org.junit.jupiter.api.MethodOrderer;
-import org.junit.jupiter.api.Order;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestInstance;
-import org.junit.jupiter.api.TestInstance.Lifecycle;
-import org.junit.jupiter.api.TestMethodOrder;
-
-import net.openchrom.wsd.converter.supplier.abif.core.ChromatogramImportConverter;
-
-@TestInstance(Lifecycle.PER_CLASS)
-@TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class AbiPrism3100GeneticAnalyser_ITest {
-
-	private IChromatogramWSD chromatogram;
-
-	@Test
-	@Order(1)
-	public void testImport() {
-
-		File fileImport = new File("testdata/files/import/3100.ab1");
-		ChromatogramImportConverter importConverter = new ChromatogramImportConverter();
-		IProcessingInfo<IChromatogramWSD> processingInfo = importConverter.convert(fileImport, new NullProgressMonitor());
-		chromatogram = processingInfo.getProcessingResult();
-		assertNotNull(chromatogram);
-	}
-
-	@Test
-	public void testWavelengths() {
-
-		IScanWSD scan = chromatogram.getScan(1);
-		assertEquals(540.0f, scan.getScanSignal(0).getWavelength(), 0);
-	}
-}
